ece417_20fall_mp6/requirements.txt
gradescope-utils>=0.3.1
numpy
pytorch>=1.7.0
tqdm
scikit-video